+++ CHANGELOG.md
@@ -31,7 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Allowed passing labels in (n_samples, n_classes) to `AveragePrecision` ([#386](https://github.com/PyTorchLightning/metrics/issues/386))
 
-- Added calibration error metrics ([#218])(https://github.com/PyTorchLightning/metrics/issues/218)
+- Added calibration error metrics ([#218](https://github.com/PyTorchLightning/metrics/issues/218))
 
 
 - Added support for negative targets in `nDCG` metric ([#378](https://github.com/PyTorchLightning/metrics/pull/378))